python/yellowstone-fumarole-client/pyproject.toml
[tool.poetry]
name = "yellowstone-fumarole-client"
version = "0.2.2"
homepage = "https://github.com/rpcpool/yellowstone-fumarole"
repository = "https://github.com/rpcpool/yellowstone-fumarole"
description = "Yellowstone Fumarole Python Client"
authors = ["Louis-Vincent <louis-vincent@triton.one>", "Triton One <help@triton.one>"]
readme = "README.md"

packages = [
    { include = "yellowstone_fumarole_proto" },
    { include = "yellowstone_fumarole_client" },
]

[tool.poetry.dependencies]
python = "^3.13"
grpcio = "^1.71.1"
protobuf = "^6.32.0"
pyyaml = "^6.0.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.68.1"
black = "^24.10.0"
pytest-asyncio = "^0.26.0"
deptry = "^0.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
